Python/DarkiDewController/requirements.txt
pyserial>=3.5
PyQt6>=6.5.0
pyqtgraph>=0.13.3
pandas>=2.0.0
numpy>=1.24.0